requirements.txt
# =========================================
# Dependências Core (Produção)
# =========================================

# Framework Web
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9,<1.0.0

# Google Gemini / ADK
google-genai>=0.2.0
google-adk>=1.18.0

# Configuração e Ambiente
python-dotenv>=1.0.0,<2.0.0
pydantic-settings>=2.0.0,<3.0.0

# Rate Limiting e Segurança
slowapi>=0.1.9,<1.0.0

# Retry e Resiliência
# Nota: google-adk>=1.18.0 requer tenacity>=9.0.0
tenacity>=9.0.0,<10.0.0

# Cache
cachetools>=5.3.0,<7.0.0

# HTTP Client (para OpenRouter)
httpx>=0.27.0,<1.0.0

# =========================================
# Dependências de Desenvolvimento
# =========================================

# Formatação e Linting
black>=24.0.0,<26.0.0
ruff>=0.5.0,<1.0.0
mypy>=1.8.0,<2.0.0

# Pre-commit Hooks
pre-commit>=3.5.0,<4.0.0

# =========================================
# Dependências de Testes
# =========================================

# Framework de Testes
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
nest_asyncio>=1.6.0,<2.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-html>=4.1.0,<5.0.0

# Testes E2E e Acessibilidade
playwright==1.51.0
pytest-playwright==0.7.1
pytest-base-url==2.1.0
axe-playwright-python==0.1.5

# Testes de Carga e Performance
locust==2.42.5
pytest-benchmark==5.2.0

# Property-Based Testing
hypothesis==6.130.0

# Mutation Testing
mutmut==3.4.0

# Relatórios Visuais
allure-pytest==2.13.5
